backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
opencv-python
numpy
Pillow
torch>=1.13
torchvision
timm>=0.6.12
mediapipe
librosa
loguru
python-dotenv
